requirements-pi.txt
# ============================================================
# Empathetic AI Companion — Raspberry Pi 5 Dependencies
# ============================================================
# Run these apt commands FIRST before pip installing:
#
#   sudo apt update && sudo apt upgrade -y
#   sudo apt install -y \
#       python3-pip python3-venv \
#       espeak-ng libespeak-ng1 \
#       libatlas-base-dev libhdf5-dev \
#       libgtk-3-dev libopencv-dev \
#       portaudio19-dev python3-pyaudio \
#       libjpeg-dev libpng-dev libtiff-dev \
#       libcamera-apps
#
# Then:
#   python3 -m venv venv
#   source venv/bin/activate
#   pip install -r requirements-pi.txt
# ============================================================

# --- Computer Vision ---
opencv-python>=4.8.0            # Full build — needed for cv2.imshow with monitor

# --- Emotion Detection ---
fer>=22.5.1                     # Pulls in tensorflow automatically

# --- Face Geometry ---
mediapipe>=0.10.9               # Tasks API FaceLandmarker

# --- Speech Recognition ---
SpeechRecognition>=3.10
PyAudio>=0.2.14

# --- TTS ---
pyttsx3>=2.90                   # Uses espeak-ng backend on Pi

# --- Groq AI (LLM + Whisper STT) ---
groq>=0.9.0
python-dotenv>=1.0.0

# --- Numerical ---
numpy>=1.24
